seldon/requirements.txt
seldon-core==1.4.0
tensorflow==2.3.1
Pillow==8.0.1
